requirements.txt
gradio
requests
python-dotenv
deep-translator
google-generativeai
pandas
pyperclip
